// src/app/core/models/perfil.enum.ts
// Enum que define os perfis de acesso do sistema de gestao de O.S.
export enum Perfil {
  // Usuario solicitante de ordens de servico.
  SOLICITANTE = 'SOLICITANTE',
  // Tecnico responsavel pela execucao das ordens de servico.
  TECNICO = 'TÉCNICO',
  // Supervisor responsavel pela gestao e aprovacao das O.S.
  SUPERVISOR = 'SUPERVISOR',
}